+++ CHANGELOG.md
@@ -1,4 +1,5 @@
 ## Version 2.4.1
+* Scrolls of new spells should now be available at appropriate venders.
 * Added Spells
     * Stunning Barrier Greater
         * This spell functions as stunning barrier, except it provides a +2 bonus to AC and on saving throws and it is not discharged until it has stunned a number of creatures equal to your caster level.